worker: improve code organization and error handling
